requirements.txt
cfgv==3.4.0
distlib==0.3.8
filelock==3.13.1
identify==2.5.33
iniconfig==2.0.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
pytest==7.4.4
PyYAML==6.0.1
setuptools==69.0.3
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
virtualenv==20.25.0
